refactor(clawdstrike): extract types and guard_configs from policy.rs
Pure relocation:
- types: Policy (+ Default), MergeStrategy, PolicyValidationOptions,
  PolicyLoadVerificationInput, private default_version
- guard_configs: GuardConfigs struct only (13 guard-config wrappers + the
  three #[serde(skip)] presence sidecars + cfg-gated spider_sense variants).
  GuardConfigs::merge_with stays in mod.rs for now (moves to merge.rs next).

mod.rs keeps the load-verifier machinery (PolicyLoadVerifier, the static, and
install_policy_load_verifier) and re-exports the new public types explicitly.
Trimmed now-unused crate::guards/posture imports from mod.rs. No logic edits.
